+++ Tests/CodableWrappersTests/Encoder/DateEncodingTests.swift
@@ -40,7 +40,7 @@ class DateEncodingTests: QuickSpec, EncodingTestSpec {
                     }
                 }
                 // MARK: ISO8601
-                if #available(OSX 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *) {
+                if #available(macOS 12, iOS 15.0, watchOS 8, tvOS 15.0, *) {
                     it("ISO8601") {
                         expect {_ = try self.jsonEncoder.encode(iso8601TestInstance)}.toNot(throwError())
                         let encodedData = try? self.jsonEncoder.encode(iso8601TestInstance)
@@ -106,7 +106,7 @@ class DateEncodingTests: QuickSpec, EncodingTestSpec {
                     }
                 }
                 // MARK: ISO8601
-                if #available(OSX 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *) {
+                if #available(macOS 12, iOS 15.0, watchOS 8, tvOS 15.0, *) {
                     it("ISO8601") {
                         expect {_ = try self.plistEncoder.encode(iso8601TestInstance)}.toNot(throwError())
                         let encodedData = try? self.plistEncoder.encode(iso8601TestInstance)
@@ -218,17 +218,21 @@ private let millisecondsSince1970XML2 = """
 """
 
 // MARK: - ISO8601 Mock Data
-@available(macOS 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *)
+@available(OSX 12, iOS 15.0, watchOS 8, tvOS 15.0, *)
 private struct ISO8601TestModel: Codable, Equatable {
     @CodingUses<ISO8601DateStaticCoder>
     var iso8601Date: Date
+    @CodingUses<ISO8601StaticCoder>
+    var iso8601DateStyle: Date
 }
-@available(macOS 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *)
-private let iso8601TestInstance = ISO8601TestModel(iso8601Date: ISO8601DateFormatter().date(from: "2008-09-15T10:53:00Z")!)
+@available(macOS 12, iOS 15.0, watchOS 8, tvOS 15.0, *)
+private let iso8601TestInstance = ISO8601TestModel(iso8601Date: ISO8601DateFormatter().date(from: "2008-09-15T10:53:00Z")!,
+                                                   iso8601DateStyle: ISO8601DateFormatter().date(from: "2008-09-15T10:53:00Z")!)
 @available(macOS 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *)
 private let iso8601JSON = """
 {
-    "iso8601Date" : "2008-09-15T10:53:00Z"
+    "iso8601Date" : "2008-09-15T10:53:00Z",
+    "iso8601DateStyle" : "2008-09-15T10:53:00Z"
 }
 """
 @available(macOS 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *)
@@ -238,6 +242,8 @@ private let iso8601XML = """
 <plist version="1.0">
 <dict>
     <key>iso8601Date</key>
+    <string>2008-09-15T10:53:00Z</string>
+    <key>iso8601DateStyle</key>
     <string>2008-09-15T10:53:00Z</string>
 </dict>
 </plist>
@@ -276,16 +282,20 @@ private struct TestCustomDateFormatter: DateFormatterStaticCoder {
 
 // MARK: - Custom ISO8601 Mock Data
 
-@available(macOS 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *)
+@available(OSX 12, iOS 15.0, watchOS 8, tvOS 15.0, *)
 private struct CustomISO8601FormatterTestModel: Codable, Equatable {
     @CodingUses<TestCustomISO8601DateFormatter>
     var customFormatDate: Date
+    @CodingUses<TestCustomISO8601DateStyle>
+    var customFormatDateStyle: Date
 }
-@available(macOS 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *)
-private let iso8601customFormatterTestInstance = CustomISO8601FormatterTestModel(customFormatDate: TestCustomISO8601DateFormatter.iso8601DateFormatter.date(from: "2011-06-10T20:24:16Z")!)
+@available(macOS 12, iOS 15.0, watchOS 8, tvOS 15.0, *)
+private let iso8601customFormatterTestInstance = CustomISO8601FormatterTestModel(customFormatDate: TestCustomISO8601DateFormatter.iso8601DateFormatter.date(from: "2011-06-10T20:24:16Z")!,
+                                                                                 customFormatDateStyle: TestCustomISO8601DateFormatter.iso8601DateFormatter.date(from: "2011-06-10T20:24:16Z")!)
 private let iso8601customFormatterJSON = """
 {
-    "customFormatDate" : "2011-06-10T20:24:16Z"
+    "customFormatDate" : "2011-06-10T20:24:16Z",
+    "customFormatDateStyle" : "20110610T20:24:16Z"
 }
 """
 private let iso8601customFormatterXML = """
@@ -295,6 +305,8 @@ private let iso8601customFormatterXML = """
 <dict>
     <key>customFormatDate</key>
     <string>2011-06-10T20:24:16Z</string>
+    <key>customFormatDateStyle</key>
+    <string>20110610T20:24:16Z</string>
 </dict>
 </plist>
 """
@@ -302,9 +314,27 @@ private let iso8601customFormatterXML = """
 // MARK: - Custom ISO8601 Formatter
 @available(macOS 10.12, iOS 10.0, watchOS 3.0, tvOS 10.0, *)
 private struct TestCustomISO8601DateFormatter: ISO8601DateFormatterStaticCoder {
+#if swift(>=6.0.0)
+    nonisolated(unsafe) static let iso8601DateFormatter: ISO8601DateFormatter = {
+        let formatter = ISO8601DateFormatter()
+        formatter.formatOptions = [.withInternetDateTime, .withDashSeparatorInDate]
+        return formatter
+    }()
+#else
     static let iso8601DateFormatter: ISO8601DateFormatter = {
         let formatter = ISO8601DateFormatter()
         formatter.formatOptions = [.withInternetDateTime, .withDashSeparatorInDate]
         return formatter
     }()
+#endif
+}
+
+// MARK: - Custom ISO8601 Date Style Formatter
+@available(swift, introduced: 5.4) @available(macOS, introduced: 12)
+@available(iOS, introduced: 15.0) @available(tvOS, introduced: 15.0)
+@available(watchOS, introduced: 8.0)
+private struct TestCustomISO8601DateStyle: ISO8601DateFormatStyleStaticCoder {
+    static var iso8601DateStyle: Date.ISO8601FormatStyle {
+        .init().dateSeparator(.omitted)
+    }
 }